flaskServer/requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.2.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.2.0
six==1.16.0
SQLAlchemy==1.4.39
typing-extensions==4.3.0
Werkzeug==2.2.1
zipp==3.8.1
